refactor: reorganize auth components structure by consolidating SignInPage, SignUpPage, and VerifyPage into a unified directory
