--- src/pages/auth/sign-in/index.ts
@@ -1,1 +1,0 @@
-export { SignInPage } from './ui/SignInPage'
