--- src/pages/auth/sign-up/index.ts
@@ -1,1 +1,0 @@
-export { SignUpPage } from './ui/SignUpPage'
